refactored chat tab file to export component
